refactor: Modifies Coverage and Accuracy Definition in TopDownMonitor.
The prefetch bit is saved when the prefetch block is accessed, using (prefetch & !accessed) to determine the first access to the prefetch block.

Coverage is defined as the number of dmandRequests that hit a prefetch block devided by the number of all demandRequests.

Accuracy defines the number of useful prefetching request sented divided by all prefetching requests